+++ bu-ussd/enrollment/translations/index.js
@@ -1,0 +1,67 @@
+module.exports = {
+    'bundles_title': {
+        'en_bu': 'Select a Product for $firstName $lastName',
+        'bu': 'Hitamwo ivyasabwe vya $firstName $lastName',
+        'fr_bu': 'Selectionnez un produit pour $firstName $lastName'
+    },
+    'next_screen': {
+        'en_bu': '77) continue',
+        'bu': '77) bandanya',
+        'fr_bu': '77) continuer'
+    },
+    'enter_quantity': {
+        'en_bu': 'Select Quantity: $bundleInput /$unit',
+        'bu': 'Hitamwo igitigiri: $bundleInput /$unit',
+        'fr_bu': 'Selectionner la quantite: $bundleInput /$unit'
+    },
+    'order_or_finalize': {
+        'en_bu': '1) Continue ordering\n2) Confirm and finalize',
+        'bu': '1) Bandanya gusaba inyongeramwimbu\n2) Kwemeza no gusozera',
+        'fr_bu': '1) Continuer a commander\n2) Confirmer et finaliser'
+    },
+    'finalize': {
+        'en_bu': '$firstName $lastName ordered for \n$order',
+        'bu': '$firstName $lastName ivyasabwe vy\' \n$order',
+        'fr_bu': '$firstName $lastName commandes pour \n$order'
+    },
+    'finalize_order': {
+        'en_bu': '$firstName $lastName ordered for \n$order\n$menu',
+        'bu': '$firstName $lastName ivyasabwe vy\' \n$order\n$menu',
+        'fr_bu': '$firstName $lastName commandes pour \n$order\n$menu'
+    },
+    'finalize_menu': {
+        'en_bu': '1) Add another product\n2) Change Order\n3) Confirm',
+        'bu': '1) Ongerako ikindi gikoresho\n2) Guhindura ivyasabwe\n3) Kwemeza',
+        'fr_bu': '1) Ajouter un autre produit\n2) Modifier la commande\n3) Confirmer'
+    },
+    'total_credit': {
+        'en_bu': 'Total Credit = $amount\n',
+        'bu': 'Ideni ryose = $amount\n',
+        'fr_bu': 'Total des credits = $amount\n'
+    },
+    'change_order_title': {
+        'en_bu': 'Remove Product from List\n0) Confirm\n',
+        'bu': 'Gukura ivyasabwe kurutonde\n0) Kwemeza\n',
+        'fr_bu': 'Retirer un produit de la liste\n0) Confirmer\n'
+    },
+    'error_enrolling': {
+        'en_bu': 'There was an error please try again/later',
+        'bu': 'Habaye ikosa, Musabwe kugerageza kandi hanyuma',
+        'fr_bu': 'une erreur s\'est produite. Reessayez plus tard'
+    },
+    'order': {
+        'en_bu': 'Order\n',
+        'bu': 'commander\n',
+        'fr_bu': 'Ivyasabwe\n'
+    },
+    'already_enrolled': {
+        'en_bu': 'This farmer has already placed an order. If you wish to change please call the call center',
+        'bu': 'Muradutunge, igihe co gusaba inyongeramwimbu cararenze. Mwipfuza guhindura, mwohamagara ku nimero itariha. Murakoze',
+        'fr_bu': 'Desole, le delai pour passer les commandes est expire. Pour modifier votre commande, merci de contacter le centre d\'appel'
+    },
+    'sample': {
+        'en_bu': '',
+        'bu': '',
+        'fr_bu': ''
+    }
+};